# packages/contracts/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/eagle/assets": {
      "get": {
        "operationId": "EagleController_listAssets",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "maximum": 100,
              "default": 30,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "maxLength": 200,
              "type": "string"
            }
          },
          {
            "name": "rating",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 5,
              "type": "number"
            }
          },
          {
            "name": "format",
            "required": false,
            "in": "query",
            "schema": {
              "maxLength": 32,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": ["eagle"]
      }
    },
    "/api/eagle/assets/{assetId}": {
      "get": {
        "operationId": "EagleController_getAsset",
        "parameters": [
          {
            "name": "assetId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": ["eagle"]
      },
      "patch": {
        "operationId": "EagleController_updateAsset",
        "parameters": [
          {
            "name": "assetId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEagleAssetDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": ["eagle"]
      }
    },
    "/api/eagle/trash/{assetId}": {
      "get": {
        "operationId": "EagleController_getTrashAsset",
        "parameters": [
          {
            "name": "assetId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": ["eagle"]
      }
    },
    "/api/eagle/assets/{assetId}/original": {
      "get": {
        "operationId": "EagleController_getOriginal",
        "parameters": [
          {
            "name": "assetId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": ["eagle"]
      }
    },
    "/api/eagle/assets/{assetId}/renditions/{renditionId}": {
      "get": {
        "operationId": "EagleController_getRendition",
        "parameters": [
          {
            "name": "assetId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "renditionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": ["eagle"]
      }
    },
    "/api/eagle/assets/trash": {
      "post": {
        "operationId": "EagleController_trashAssets",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EagleAssetIdsDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": ["eagle"]
      }
    },
    "/api/eagle/trash": {
      "get": {
        "operationId": "EagleController_listTrash",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "maximum": 100,
              "default": 30,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "maxLength": 200,
              "type": "string"
            }
          },
          {
            "name": "rating",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 5,
              "type": "number"
            }
          },
          {
            "name": "format",
            "required": false,
            "in": "query",
            "schema": {
              "maxLength": 32,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": ["eagle"]
      },
      "delete": {
        "operationId": "EagleController_emptyTrash",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": ["eagle"]
      }
    },
    "/api/eagle/trash/restore": {
      "post": {
        "operationId": "EagleController_restoreAssets",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EagleAssetIdsDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": ["eagle"]
      }
    },
    "/api/eagle/tags": {
      "get": {
        "operationId": "EagleController_listTags",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": ["eagle"]
      },
      "post": {
        "operationId": "EagleController_createTag",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateManualTagDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": ["eagle"]
      }
    },
    "/api/eagle/tags/{tagId}": {
      "patch": {
        "operationId": "EagleController_updateTag",
        "parameters": [
          {
            "name": "tagId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateManualTagDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": ["eagle"]
      },
      "delete": {
        "operationId": "EagleController_deleteTag",
        "parameters": [
          {
            "name": "tagId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": ["eagle"]
      }
    },
    "/api/eagle/tag-groups": {
      "get": {
        "operationId": "EagleController_listTagGroups",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": ["eagle"]
      },
      "post": {
        "operationId": "EagleController_createTagGroup",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateManualTagGroupDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": ["eagle"]
      }
    },
    "/api/eagle/tag-groups/{groupId}": {
      "patch": {
        "operationId": "EagleController_updateTagGroup",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateManualTagGroupDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": ["eagle"]
      },
      "delete": {
        "operationId": "EagleController_deleteTagGroup",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": ["eagle"]
      }
    },
    "/api/eagle/ai-tags": {
      "get": {
        "operationId": "EagleController_listAiTags",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": ["eagle"]
      }
    },
    "/api/eagle/assets/{assetId}/tags": {
      "put": {
        "operationId": "EagleController_replaceAssetTags",
        "parameters": [
          {
            "name": "assetId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplaceAssetTagsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": ["eagle"]
      }
    },
    "/api/eagle/smart-folders": {
      "get": {
        "operationId": "EagleController_listSmartFolders",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": ["eagle"]
      },
      "post": {
        "operationId": "EagleController_createSmartFolder",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSmartFolderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": ["eagle"]
      }
    },
    "/api/eagle/smart-folders/{folderId}": {
      "patch": {
        "operationId": "EagleController_updateSmartFolder",
        "parameters": [
          {
            "name": "folderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSmartFolderDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": ["eagle"]
      },
      "delete": {
        "operationId": "EagleController_deleteSmartFolder",
        "parameters": [
          {
            "name": "folderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": ["eagle"]
      }
    },
    "/api/eagle/smart-folders/{folderId}/move": {
      "post": {
        "operationId": "EagleController_moveSmartFolder",
        "parameters": [
          {
            "name": "folderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MoveSmartFolderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": ["eagle"]
      }
    },
...
  },
  "info": {
    "title": "SekerEagle API",
    "description": "",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "Object": {
        "type": "object",
        "properties": {}
      },
      "UpdateEagleAssetDto": {
        "type": "object",
        "properties": {
          "rowVersion": {
            "type": "number",
            "minimum": 1
          },
          "displayName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 255
          },
          "rating": {
            "type": "object",
            "nullable": true,
            "minimum": 1,
            "maximum": 5
          },
          "color": {
            "type": "object",
            "nullable": true,
            "pattern": "^#[0-9a-fA-F]{6}$"
          },
          "description": {
            "type": "object",
            "nullable": true,
            "maxLength": 4000
          },
          "sourceUrl": {
            "type": "object",
            "nullable": true,
            "maxLength": 2048
          }
        },
        "required": ["rowVersion"]
      },
      "EagleAssetIdsDto": {
        "type": "object",
        "properties": {
          "assetIds": {
            "maxItems": 100,
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": ["assetIds"]
      },
      "CreateManualTagDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "color": {
            "type": "object",
            "nullable": true,
            "pattern": "^#[0-9a-fA-F]{6}$"
          }
        },
        "required": ["name"]
      },
      "UpdateManualTagDto": {
        "type": "object",
        "properties": {
          "rowVersion": {
            "type": "number",
            "minimum": 1
          }
        },
        "required": ["rowVersion"]
      },
      "CreateManualTagGroupDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateManualTagGroupDto": {
        "type": "object",
        "properties": {}
      },
      "ReplaceAssetTagsDto": {
        "type": "object",
        "properties": {
          "tagIds": {
            "maxItems": 100,
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": ["tagIds"]
      },
      "CreateSmartFolderDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "parentId": {
            "type": "object",
            "nullable": true
          },
          "color": {
            "type": "object",
            "nullable": true,
            "pattern": "^#[0-9a-fA-F]{6}$"
          },
          "query": {
            "type": "object",
            "additionalProperties": true
          }
        },
        "required": ["name", "query"]
      },
      "UpdateSmartFolderDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "parentId": {
            "type": "object",
            "nullable": true
          },
          "color": {
            "type": "object",
            "nullable": true,
            "pattern": "^#[0-9a-fA-F]{6}$"
          },
          "query": {
            "type": "object",
            "additionalProperties": true
          },
          "rowVersion": {
            "type": "number",
            "minimum": 1
          }
        },
        "required": ["name", "query", "rowVersion"]
      },
      "MoveSmartFolderDto": {
        "type": "object",
        "properties": {
          "rowVersion": {
            "type": "number",
            "minimum": 1
          },
          "parentId": {
            "type": "object",
            "nullable": true
          },
          "position": {
            "type": "number",
            "minimum": 0
          }
        },
        "required": ["rowVersion", "position"]
      },
...
    }
  }
}